pre compute gulmc structure
